Initial commit of new LaTeX template (see notes)
I have decided to create this new repository for the following reason: it
allows to simply clone this repo into a project; with the other one
(Preperation) you'd have to remove all other folders and files.

Additionally, this new template is a combination of both the Basic- and
the Advanced report, as well as some other templates. This new template is
really simple, but at the moment not quite finished (needs some finishing
touches).

The shell and batch script have also been modified - although the batch
file isn't quite as good as the shell script since writing batch scritps
is dumb.

This repository should now offer everything needed to write good reports,
thesis, protocols or whatever you may need it to be.